backend_django/requirements.txt
django>=5.0,<6.0
djangorestframework>=3.15,<3.16
djangorestframework-simplejwt>=5.3,<6.0
python-dotenv>=1.0,<2.0
django-cors-headers>=4.3,<5.0
gunicorn>=22,<24
whitenoise>=6,<7
dj-database-url>=2,<3
psycopg2-binary>=2.9,<3
